+++ src/generate/hightMapGen/terrainGen.hpp
@@ -4,5 +4,5 @@
 #include <vector>
 
 namespace HeightMapGen {
-  std::vector<uint8_t> GenerateNewHeightMap(uint32_t seed, int8_t width, uint8_t height);
+  std::vector<uint8_t> GenerateNewHeightMap(uint32_t seed, int32_t width, uint32_t height);
 }
